+++ .idea/.gitignore 2
@@ -1,0 +1,3 @@
+# Default ignored files
+/shelf/
+/workspace.xml
