+++ .idea/.name 2
@@ -1,0 +1,1 @@
+Movie App
