+++ app/.gitignore 2
@@ -1,0 +1,1 @@
+/build
