+++ app/src/main/java/com/example/movieapp/MainActivity.kt
@@ -26,7 +26,7 @@ class MainActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications
+                R.id.navigation_home, R.id.navigation_favorite, R.id.navigation_details
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
