+++ app/src/main/java/com/example/movieapp/ui/favorite/FavoriteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.movieapp.ui.dashboard
+package com.example.movieapp.ui.favorite
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.movieapp.databinding.FragmentDashboardBinding
+import com.example.movieapp.databinding.FragmentFavoriteBinding
 
-class DashboardFragment : Fragment() {
+class FavoriteFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentFavoriteBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val favoriteViewModel =
+            ViewModelProvider(this).get(FavoriteViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentFavoriteBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textFavorite
+        favoriteViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
