+++ app/src/main/java/com/example/movieapp/ui/favorite/FavoriteViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.movieapp.ui.notifications
+package com.example.movieapp.ui.favorite
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class FavoriteViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is favorite Fragment"
     }
     val text: LiveData<String> = _text
 }
