+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,13 +7,13 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
+        android:id="@+id/navigation_favorite"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_favorite" />
 
     <item
-        android:id="@+id/navigation_notifications"
+        android:id="@+id/navigation_details"
         android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:title="@string/title_details" />
 
 </menu>
